chore: make a generic map render
