--- packages/visualizations/src/components/MapPoi/mapStyles.ts
@@ -1,9 +1,0 @@
-import type { StyleSpecification } from 'maplibre-gl';
-
-// eslint-disable-next-line import/prefer-default-export
-export const BLANK: StyleSpecification = {
-    version: 8,
-    sources: {},
-    metadata: {},
-    layers: [],
-};
