+++ packages/visualizations/src/components/utils/Map/Map.ts
@@ -29,14 +29,13 @@ import {
     POPUP_WIDTH,
     POPUP_NAVIGATION_CONTROLS_OFFSET_CLASSNAME,
     DEFAULT_SORT_KEY_VALUE,
-} from './constants';
+ POPUP_DISPLAY } from './constants';
 import type {
     PopupConfigurationByLayers,
     CenterZoomOptions,
     PopupDisplayTypes,
     Images,
 } from './types';
-import { POPUP_DISPLAY } from './types';
 
 const CURSOR = {
     DEFAULT: 'default',
