+++ packages/visualizations/src/components/utils/Map/index.ts
@@ -1,0 +1,4 @@
+import Map from "./Map.svelte";
+import 'maplibre-gl/dist/maplibre-gl.css';
+
+export default Map;
